tests/wpt/tests/html/canvas/offscreen/reset/2d.reset.state.line_join.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.reset.state.line_join</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.reset.state.line_join</h1>
<p class="desc">check that the state is reset</p>


<script>
var t = async_test("check that the state is reset");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  const default_value = ctx.lineJoin;

  ctx.lineJoin = 'bevel';
  _assert(ctx.lineJoin == 'bevel', "ctx.lineJoin == 'bevel'");

  ctx.reset();
  _assert(ctx.lineJoin == default_value, "ctx.lineJoin == default_value");
  t.done();

});
</script>
